距离boos关卡图标提示；金币数量上限；货币单位转换加小数

+++ UnityMiniGameFramework/ChickenMaster/Data/LocalUserData.cs
@@ -108,7 +108,7 @@ namespace UnityMiniGameFramework
 
     public class LocalBaseInfo
     {
-        public int gold { get; set; }
+        public long gold { get; set; }
         public int level { get; set; }
         public int exp { get; set; }
